requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.2.10
django-allauth==0.48.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
hyperlink==21.0.0
incremental==21.3.0
oauthlib==3.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
zope.interface==5.4.0
